+++ Scripts/Items/Terminal_Patches.cs
@@ -2,7 +2,9 @@
 using System.Collections.Generic;
 using System.Text;
 using HarmonyLib;
+using MikesTweaks.Scripts.Moons;
 using MikesTweaks.Scripts.World;
+using Unity.Netcode;
 
 namespace MikesTweaks.Scripts.Items
 {
@@ -10,9 +12,13 @@ namespace MikesTweaks.Scripts.Items
     public class Terminal_Patches
     {
         [HarmonyPatch("Start")]
+        [HarmonyPostfix]
         public static void Start(Terminal __instance)
         {
-            
+            if (!NetworkManager.Singleton.IsServer)
+                return;
+
+            MoonTweaks.ReapplyConfigs(__instance);
         }
     }
 }